pyproject.toml
[tool.poetry]
name = "resource_monitor"
version = "0.1.0"
description = "Show information about available resources"
authors = ["grm <golosov_SA@mail.ru>"]
readme = "README.md"
packages = [{include = "resource_monitor"}]

[tool.poetry.dependencies]
python = "^3.11"
psutil = "^5.9.4"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
